Add models based on CNN+LSTM to generate or classify
These models are composed of a CNN to extract features from the original
image and, optionally, downsample it. This CNN can be classical or
quantum. Then, a LSTM or a QLSTM can be applied to generate the final
prediction. The model can be used to model the input data distribution
via language modelling (the next pixel is predicted) or to classify the
input data into the pre-determined classes of the dataset.